commits as of 07-02-22

+++ src/main/java/com/zee/zee5app/repository/UserRepository.java
@@ -1,6 +1,6 @@
 package com.zee.zee5app.repository;
 
-import java.math.BigDecimal;
+import java.math.BigInteger;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
@@ -10,5 +10,5 @@ import com.zee.zee5app.dto.Register;
 @Repository
 public interface UserRepository extends JpaRepository<Register, String> {
 	
-	boolean existsByEmailAndContactnumber(String email,BigDecimal contactnumber);
+	boolean existsByEmailAndContactnumber(String email,BigInteger contactnumber);
 }
